[Chore] KeyboardAdaptiveModifier File create (#39)
